requirements-pythonanywhere.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.4
python-dotenv==1.0.0
bcrypt==4.0.1
PyMySQL==1.1.0